[RDSS-955] Allow sleep interval to be customised
When creating the Arkivum Adapter, which reads from the Kinesis streams
shared_services_output_*, when it reached the most recent messages, it
would raise the exception

    botocore.errorfactory.ProvisionedThroughputExceededException: An
    error occurred (ProvisionedThroughputExceededException) when calling
    the GetRecords operation: Rate exceeded for shard
    shardId-000000000000 in stream shared_services_output_dev under
    account 458323522494

The interval of 0.2 does appear to respect the limit at
https://docs.aws.amazon.com/streams/latest/dev/service-sizes-and-limits.html

    Each shard can support up to 5 transactions per second for reads

However, this only holds when there is 1 client per stream: I suspect
there are now at least 2 readers of the shared_services_output stream. I
suspect Kinesis is not designed for this case: multiple independent
consumers of a stream. Some longer term architectural descisions have to
be made in this respect: and this is a short term attempt to try to
avoid the exception.