[RDSS-1929] Make decorator more robust against evil messages
Note the covention here is apparently to use the "_json" suffix for
dictionaries that were once json.

Am torn about whether this function should throw when encountering
invalid / unexpected json, or just silently continue. However, because
this code is to be used in error handlers in client code, to avoid
_itself_ throwing error, opting to make it work in every conceivable
case. If the calling code wants to avoid decoration when invalid, it has
to do it itself.